Updated view data model id from int to long

+++ app/src/main/java/es/developer/achambi/bproject/add/Product.java
@@ -17,7 +17,7 @@ public class Product implements Parcelable {
     }
 
     protected Product(Parcel in) {
-        id = in.readInt();
+        id = in.readLong();
         productName = in.readString();
         productType = in.readString();
     }
